requirements.txt
dash
plotly
dash-bootstrap-components
pandas
numpy
requests
gunicorn
langchain
langchain-openai
markdown==3.6
python-dotenv==1.0.1
openpyxl
pymongo
langgraph==0.6.6
matplotlib
kaleido
IPython
filelock>=3.13.0
faiss-cpu
python-docx


azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-core>=1.29.0



